requirements.txt
python-dotenv==1.0.0
psutil==5.9.6
requests==2.31.0
schedule==1.2.0
sqlalchemy==2.0.23
click==8.1.7 